\name{ESLCCA-package}
\alias{ESLCCA-package}
\docType{package}
\title{ Extended Semi-Linear Canonical Correlation Analysis }
\description{
ESLCCA estimates the parameters of a given nonlinear pharmaocometric  model to maximize the correlation with a linear combination of the multiple response variables (in this case, the power spectra). 
}
\details{
 This method can be used to characterize brain activity under different treatment regimens in research and development projects at Pfizer.
}
\author{
Foteini Strimenopoulou and Heather Turner

Maintainer: Foteini Strimenopoulou <Foteini.Strimenopoulou@pfizer.com>
}
\references{
  Brain, P., Strimenopoulou, F. and Ivarsson, M. (2011). Analysing
  electroencephalogram (EEG) data using Extended Semi-Linear Canonical
  Correlation Analysis.  Submitted.
}

\keyword{ package }
\seealso{
\code{\link{ESLCCA}} ~~
}
\examples{
}
