\name{clonidine}
\alias{clonidine}
\docType{data}
\title{
  Clonindine Dark data
}
\description{
data on rats
}
\usage{data(clonidine)}
\format{
  The format is:
List of 2
 $ Y   : num [1:863, 1:36] 0.000127 0.00012 0.000104 0.000104 0.000113 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:863] "1" "2" "3" "4" ...
  .. ..$ : chr [1:36] "X.0_98" "X.1_95" "X.2_93" "X.3_91" ...
 $ data:'data.frame':	863 obs. of  3 variables:
  ..$ subject  : Factor w/ 2 levels "64","65": 1 1 1 1 1 1 1 1 1 1 ...
  ..$ treatment: Factor w/ 3 levels "Control","High Dose",..: 1 1 1 1 1 1 1 1 1 1 ...
  ..$ time     : int [1:863] 0 300 600 900 1200 1500 1800 2100 2400 2700 ...
}
\details{
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(clonidine)
}
\keyword{datasets}
